@@ -3,6 +3,10 @@ import DependenciesMacros
 import Foundation
 import SharedModels
 
+public enum DataClientError: Error, Equatable {
+  case resourceNotFound(String)
+}
+
 @DependencyClient
 public struct DataClient: Sendable {
   public var fetchDay1: @Sendable (_ year: ConferenceYear) throws -> Conference
@@ -17,47 +21,41 @@ public struct DataClient: Sendable {
 extension DataClient: DependencyKey {
   static public let liveValue: DataClient = .init(
     fetchDay1: { year in
-      let data = loadDataFromBundle(fileName: "\(year.rawValue)-day1")
-      let response = try jsonDecoder.decode(Conference.self, from: data)
-      return response
+      let data = try loadDataFromBundle(fileName: "\(year.rawValue)-day1")
+      return try jsonDecoder.decode(Conference.self, from: data)
     },
     fetchDay2: { year in
-      let data = loadDataFromBundle(fileName: "\(year.rawValue)-day2")
-      let response = try jsonDecoder.decode(Conference.self, from: data)
-      return response
+      let data = try loadDataFromBundle(fileName: "\(year.rawValue)-day2")
+      return try jsonDecoder.decode(Conference.self, from: data)
     },
     fetchDay3: { year in
-      let data = loadDataFromBundle(fileName: "\(year.rawValue)-day3")
-      let response = try jsonDecoder.decode(Conference.self, from: data)
-      return response
+      let data = try loadDataFromBundle(fileName: "\(year.rawValue)-day3")
+      return try jsonDecoder.decode(Conference.self, from: data)
     },
     fetchWorkshop: { year in
-      let data = loadDataFromBundle(fileName: "\(year.rawValue)-workshop")
-      let response = try jsonDecoder.decode(Conference.self, from: data)
-      return response
+      let data = try loadDataFromBundle(fileName: "\(year.rawValue)-workshop")
+      return try jsonDecoder.decode(Conference.self, from: data)
     },
     fetchSponsors: { year in
-      let data = loadDataFromBundle(fileName: "\(year.rawValue)-sponsors")
-      let response = try jsonDecoder.decode(Sponsors.self, from: data)
-      return response
+      let data = try loadDataFromBundle(fileName: "\(year.rawValue)-sponsors")
+      return try jsonDecoder.decode(Sponsors.self, from: data)
     },
     fetchOrganizers: { year in
-      let data = loadDataFromBundle(fileName: "\(year.rawValue)-organizers")
-      let response = try jsonDecoder.decode([Organizer].self, from: data)
-      return response
+      let data = try loadDataFromBundle(fileName: "\(year.rawValue)-organizers")
+      return try jsonDecoder.decode([Organizer].self, from: data)
     },
     fetchSpeakers: { year in
-      let data = loadDataFromBundle(fileName: "\(year.rawValue)-speakers")
-      let response = try jsonDecoder.decode([Speaker].self, from: data)
-      return response
+      let data = try loadDataFromBundle(fileName: "\(year.rawValue)-speakers")
+      return try jsonDecoder.decode([Speaker].self, from: data)
     }
   )
 
-  static func loadDataFromBundle(fileName: String) -> Data {
-    let filePath = Bundle.module.path(forResource: fileName, ofType: "json")!
+  static func loadDataFromBundle(fileName: String) throws -> Data {
+    guard let filePath = Bundle.module.path(forResource: fileName, ofType: "json") else {
+      throw DataClientError.resourceNotFound(fileName)
+    }
     let fileURL = URL(fileURLWithPath: filePath)
-    let data = try! Data(contentsOf: fileURL)
-    return data
+    return try Data(contentsOf: fileURL)
   }
 }
 
